define performance.timeOrigin attribute
The returned value is the global time of the "zero time" of the time
origin. This allows multiple contexts, each with own time origin, to
translate timestamps with sub-millisecond resolution, either by
communicating their global time to the other context, or first
translating their timestamps against the global time of their time
origin. Related discussions in #21 and #22.

A polyfill for this is ~: Date.now()-performance.now(), modulo clock
skew and adjustments that Date.now() is subject to.